Add a method to write logs to the console only
By default, the `logging` module logs to both console and the `log.txt`
file. Since files are stored in flash memory on most devices running
MicroPython, writing to the log file can wear the flash unnecessarily,
especially when a large number of messages are written to the log.

Let's add a simple method to write log messages to the console only.
By setting `logging.log_file` to None, the user can completely bypass
writing log messages to the log file.

Closes #64